+++ lottery-domain/src/main/java/com/lottery/domain/strategy/service/rule/tree/factory/engine/impl/DecisionTreeEngineImpl.java
@@ -27,26 +27,27 @@ public class DecisionTreeEngineImpl implements DecisionTreeEngine {
         this.ruleTreeVO = ruleTreeVO;
     }
     @Override
-    public RuleEntity process(String userId, Long strategyId, Integer awardId) {
+    public RuleEntity process(String userId, Long strategyId, Long awardId) {
         RuleEntity ruleEntity = null;
-        // 获取基础信息（根节点、所有规则树节点的Map集合）
+        // 1. 获取基础信息（根节点、所有规则树节点的Map集合）
         String nextNode = ruleTreeVO.getTreeRootRuleNode();
         Map<String, RuleTreeNodeVO> treeNodeMap = ruleTreeVO.getTreeNodeMap();
-        // 获取起始节点「根节点记录第一个要执行的规则」
+        // 2. 获取起始节点（根节点记录第一个要执行的规则）
         RuleTreeNodeVO ruleTreeNode = treeNodeMap.get(nextNode);
+        // 3. 按条件遍历规则树节点
         while (ruleTreeNode!=null) {
-            // 获取规则树的节点
+            // 3.1. 获取规则树的节点
             LogicTree logicTreeNode = logicTreeNodeGroup.get(ruleTreeNode.getRuleName());
-            // 节点计算，判断走向
+            // 3.2. 节点计算，判断走向
             DefaultLogicTreeFactory.TreeActionEntity logicEntity = logicTreeNode.logic(userId, strategyId, awardId);
             RuleLogicCheckTypeVO ruleLogicCheckTypeVO = logicEntity.getRuleLogicCheckType();
             ruleEntity = logicEntity.getRuleEntity();
             log.info("决策树引擎【{}】treeId:{} node:{} code:{}", ruleTreeVO.getTreeName(), ruleTreeVO.getTreeId(), nextNode, ruleLogicCheckTypeVO.getCode());
-            // 获取下个节点
+            // 3.3. 获取下个节点
             nextNode = getNextNode(ruleLogicCheckTypeVO.getCode(), ruleTreeNode.getTreeNodeLineVOList());
             ruleTreeNode = treeNodeMap.get(nextNode);
         }
-        // 返回最终结果
+        // 4. 返回最终结果
         return ruleEntity;
     }
     public String getNextNode(Integer matterValue, List<RuleTreeNodeLineVO> treeNodeLineVOList) {
